+++ frontend/src/containers/App.js
@@ -4,7 +4,7 @@ import "../styles/App.scss";
 
 import UniversalLandingPage from "../containers/UniversalLandingPage.js";
 import HostMenuView from "./HostMenuView";
-import DecksView from "./DecksView";
+import MyGamesPage from "./MyGamesPage";
 
 import { Route, Switch } from "react-router-dom";
 
@@ -15,7 +15,7 @@ function App() {
         <Route path="/player" />
 
         <Route path="/host/edit/:id" />
-        <Route path="/host/decks" component={DecksView}/>
+        <Route path="/host/decks" component={MyGamesPage}/>
         <Route path="/host/room/:id" />
         <Route path="/host" component={HostMenuView} />
