+++ frontend/src/containers/MyGamesPage/components/DeckList/components/DeckCard/styles.module.scss
@@ -1,4 +1,4 @@
-@import 'colors';
+@import '../../../../../../styles/colors';
 
 .DeckCard {
   width: 96%;
