+++ frontend/src/containers/MyGamesPage/components/DeckList/styles.module.scss
@@ -1,5 +1,5 @@
-@import 'colors';
-@import 'scroller';
+@import '../../../../styles/colors';
+@import '../../../../styles/scroller';
 
 .DeckList {
   width: 55vw;
