+++ frontend/src/containers/MyGamesPage/styles.module.scss
@@ -1,5 +1,5 @@
-@import 'colors';
-@import 'fonts';
+@import '../../styles/colors';
+@import '../../styles/fonts';
 
 .DecksView {
   position: absolute;
